Added some casts and changed some variable types to eliminate clang integer type conversion warnings.
git-svn-id: https://svn.r-project.org/R/trunk@73385 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/include/Defn.h
@@ -401,10 +401,10 @@ typedef struct {
 
 /* Hashing Macros */
 #define HASHASH(x)      ((x)->sxpinfo.gp & HASHASH_MASK)
-#define HASHVALUE(x)    TRUELENGTH(x)
+#define HASHVALUE(x)    ((int) TRUELENGTH(x))
 #define SET_HASHASH(x,v) ((v) ? (((x)->sxpinfo.gp) |= HASHASH_MASK) : \
 			  (((x)->sxpinfo.gp) &= (~HASHASH_MASK)))
-#define SET_HASHVALUE(x,v) SET_TRUELENGTH(x, v)
+#define SET_HASHVALUE(x,v) SET_TRUELENGTH(x, ((int) (v)))
 
 /* Vector Heap Structure */
 typedef struct {
